Move ImplicitSpan to candidate phase
