+++ fonduer/candidates/models/__init__.py
@@ -1,3 +1,4 @@
 from fonduer.candidates.models.candidate import Candidate, candidate_subclass
+from fonduer.candidates.models.implicitspan import ImplicitSpan, TemporaryImplicitSpan
 
-__all__ = ["Candidate", "candidate_subclass"]
+__all__ = ["Candidate", "candidate_subclass", "ImplicitSpan", "TemporaryImplicitSpan"]
