+++ fonduer/features/core_features.py
@@ -1,4 +1,5 @@
-from fonduer.parser.models import ImplicitSpan, TemporarySpan
+from fonduer.candidates.models import ImplicitSpan
+from fonduer.parser.models import TemporarySpan
 
 FEAT_PRE = "CORE_"
 DEF_VALUE = 1
